// frontend/src/core/api/model/patientsControllerFindAllSortOrder.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Dentizy API
 * API Documentation untuk Sistem Manajemen Klinik Gigi
 * OpenAPI spec version: 1.0
 */

export type PatientsControllerFindAllSortOrder = typeof PatientsControllerFindAllSortOrder[keyof typeof PatientsControllerFindAllSortOrder];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PatientsControllerFindAllSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;